+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Config option for delaying entry from when the land dimension is generated, which can help with performance.
 Note that this option is found in the common config and not the world config.
 - Particle effect while entering
+- Transportalizer particle
 
 ### Fixed
 
@@ -21,11 +22,15 @@ Note that this option is found in the common config and not the world config.
 - Fixed editmode players not being able to destroy machines
 - Messages for more cases where entry may fail
 - Spawn height for re-entry is now the top motion-blocking block
+- Fixed entering editmode sometimes placing the player at the wrong height
 - Prevent an overflowed grist gutter from negatively influencing ability to pick up grist
+- Fixed grist cache view in editmode using the limit of the wrong player
+- Cache limit will now be accounted for more often client-side
+- Clearer and more accurate grist toasts when at cache capacity
 
 ### Contributors for this release
 
-- kirderf1, Caldw3ll
+- kirderf1, Caldw3ll, Dweblenod, Riotmode
 
 ## [1.19.2-1.10.0.1] - 2023-04-13